Safe defaults for limitnewcrashers, limitnewcorpus
The purpose of these defaults is to prevent fuzzing from consuming all available disk space.  This allows underutilized servers to fuzz unattended (with nice 19, etc.) with less risk of bringing down the server.

Set the default for limitnewcrashers to 10.  This is set low because it can grow very quickly.

Set the default for limitnewcorpus to 1000.  So if fuzzing starts with 500 corpus, it will exit automatically shortly after 1500 corpus is detected.